% Generated by roxygen2 (4.0.2): do not edit by hand
\name{r_versions}
\alias{r_versions}
\title{Query which versions R-release and R-oldrel refer to}
\usage{
r_versions()
}
\value{
A named character vector with entries \code{release} and
\code{oldrel}.
}
\description{
R version numbers consist of three numbers (since version 1.4.1):
major, minor and patch.
}
\details{
We use the latest (in terms of versions numbers) tag in the
R SVN repository to determine R-release.

For R-oldrel, the latest version of the previous minor version
is used, again, based on the SVN repo.
}
\examples{
r_versions()
}

