% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rversions-package.R
\docType{package}
\name{rversions-package}
\alias{rversions}
\alias{rversions-package}
\title{rversions: Query 'R' Versions, Including 'r-release' and 'r-oldrel'}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Query the main 'R' 'SVN' repository to find the
    versions 'r-release' and 'r-oldrel' refer to, and also all previous
    'R' versions and their release dates.
}
\details{
The R version numbers and dates are extracted from the
main R SVN repository at \url{http://svn.r-project.org/R/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-hub/rversions}
  \item \url{https://r-hub.github.io/rversions}
  \item Report bugs at \url{https://github.com/r-hub/rversions/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item Jeroen Ooms \email{jeroen.ooms@stat.ucla.edu} [contributor]
  \item R Consortium [funder]
}

}
\keyword{internal}
