% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rversions.R
\name{r_release}
\alias{r_release}
\title{Version number of R-release}
\usage{
r_release(dots = TRUE)
}
\arguments{
\item{dots}{Whether to use dots instead of dashes in the version
number.}
}
\value{
A one row data frame, with columns \sQuote{version} and
  \sQuote{date}.
}
\description{
The latest tag in the SVN repository (in terms of version numbers,
not dates).
}
\examples{
r_release()
}

