% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macos.R
\name{r_release_macos}
\alias{r_release_macos}
\title{Latest R release with a macOS installer available for download}
\usage{
r_release_macos()
}
\value{
A one row data frame, with columns \sQuote{version},
\sQuote{date}, \sQuote{URL} and \sQuote{nickname}.
}
\description{
\code{\link[=r_release]{r_release()}} works based on the SVN tags in the main R source code
repository. However, an SVN tag does not mean that the same version
is available for download. \code{r_release_macos} returns the latest
version for which a macOS installer is available.
}
\examples{
r_release()
r_release_macos()
}
\seealso{
Other R downloads: \code{\link{r_release_tarball}},
  \code{\link{r_release_win}}
}
\concept{R downloads}
