% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win.R
\name{r_release_win}
\alias{r_release_win}
\title{Latest R release with a Windows installer available for download}
\usage{
r_release_win()
}
\value{
A one row data frame, with columns \sQuote{version},
\sQuote{date}, \sQuote{URL} and \sQuote{nickname}.
}
\description{
\code{\link[=r_release]{r_release()}} works based on the SVN tags in the main R source code
repository. However, an SVN tag does not mean that the same version
is available for download. \code{r_release_win} returns the latest
version for which a Windows installer is available.
}
\examples{
r_release()
r_release_win()
}
\seealso{
Other R downloads: 
\code{\link{r_release_macos}()},
\code{\link{r_release_tarball}()}
}
\concept{R downloads}
