% Generated by roxygen2 (4.0.2): do not edit by hand
\name{r_versions}
\alias{r_versions}
\title{Query R's past and present versions}
\usage{
r_versions(dots = TRUE)
}
\arguments{
\item{dots}{Whether to use dots instead of dashes in the version
number.}
}
\value{
A list, where each entry is an R version and has two
  entries: \sQuote{version} is the version number itself,
  \sQuote{date} is the exact date and time when this R version was
  released.
}
\description{
R version numbers consist of three numbers (since version 1.4.1):
major, minor and patch.
}
\details{
We extract the version numbers from the tags in the SVN repository.
}
\examples{
r_versions()
}

