% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rversions.R
\name{r_versions}
\alias{r_versions}
\title{Query R's past and present versions}
\usage{
r_versions(dots = TRUE)
}
\arguments{
\item{dots}{Whether to use dots instead of dashes in the version
number.}
}
\value{
A data frame with three columns: \sQuote{version},
\sQuote{date} and \sQuote{nickname}.
}
\description{
R version numbers consist of three numbers (since version 1.4.1):
major, minor and patch.
}
\details{
We extract the version numbers from the tags in the SVN repository.
}
\examples{
\dontshow{if (rversions:::is_online()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
r_versions()
\dontshow{\}) # examplesIf}
}
