% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rversions.R
\name{r_versions}
\alias{r_versions}
\title{Query R's past and present versions}
\usage{
r_versions(dots = TRUE)
}
\arguments{
\item{dots}{Whether to use dots instead of dashes in the version
number.}
}
\value{
A data frame with two chracter columns: \sQuote{version} and
  \sQuote{date}.
}
\description{
R version numbers consist of three numbers (since version 1.4.1):
major, minor and patch.
}
\details{
We extract the version numbers from the tags in the SVN repository.
}
\examples{
r_versions()
}

